lapack/testlapack: move randomOrthogonal
